services/requirements-translation.txt
# Python dependencies for translation service
azure-cognitiveservices-speech==1.35.0
